core(alert) use alert.tsx convention
